ディレクトリ構造の整理, またshow-detail.htmlの修正
